Move complex config fields to separate package
