+++ shortcut_composer/config_system/fields/__init__.py
@@ -1,0 +1,4 @@
+from .dual_field import DualField
+from .field_with_editable_default import FieldWithEditableDefault
+
+__all__ = ["DualField", "FieldWithEditableDefault"]
