+++ shortcut_composer/templates/pie_menu_utils/pie_config.py
@@ -2,128 +2,14 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from abc import ABC, abstractmethod
-from typing import List, Generic, TypeVar, Optional, Union, Callable
+from typing import List, Generic, TypeVar, Optional, Union
 from PyQt5.QtGui import QColor
 from config_system import Field, FieldGroup
+from config_system.fields import DualField, FieldWithEditableDefault
 from data_components import Tag
 
 T = TypeVar("T")
-F = TypeVar("F", bound=Field)
 U = TypeVar("U")
-
-
-class DualField(Field, Generic[T]):
-    """
-    Field switching save location based on passed field.
-
-    Implementation uses two identical fields, but with different save
-    location. Each time DualField is red or written, correct field is
-    picked from the determiner field.
-
-    NOTE: Callbacks are always stored in the global field, as they
-    wouldn't run in local one when switching between documents.
-    """
-    def __new__(cls, *args, **kwargs) -> 'DualField[T]':
-        obj = object.__new__(cls)
-        obj.__init__(*args, **kwargs)
-        return obj
-
-    def __init__(
-        self,
-        group: FieldGroup,
-        is_local_determiner: Field[bool],
-        field_name: str,
-        default: T,
-        parser_type: Optional[type] = None
-    ) -> None:
-        self.name = field_name
-        self.config_group = group.name
-        self._is_local_determiner = is_local_determiner
-        self._is_local_determiner.register_callback(self.refresh)
-        self._loc = group.field(field_name, default, parser_type, local=True)
-        self._glob = group.field(field_name, default, parser_type, local=False)
-
-    @property
-    def default(self):
-        return self._glob.default
-
-    @default.setter
-    def default(self, value: T):
-        self._loc.default = value
-        self._glob.default = value
-
-    def write(self, value: T) -> None:
-        """
-        Write to correct internal fields, based on determiner.
-
-        Global field must always be written to activate callbacks.
-        """
-        if self._is_local_determiner.read():
-            self._loc.write(value)
-        self._glob.write(value)
-
-    def read(self) -> T:
-        """Read from local or global field, based on determiner."""
-        if self._is_local_determiner.read():
-            return self._loc.read()
-        return self._glob.read()
-
-    def register_callback(self, callback: Callable[[], None]) -> None:
-        """Subscribe callback to both fields, as only one changes on write."""
-        self._glob.register_callback(callback)
-
-    def reset_default(self) -> None:
-        """Reset both fields to default."""
-        self._loc.reset_default()
-        self._glob.reset_default()
-
-    def refresh(self) -> None:
-        """
-        Write red value back to itself.
-
-        Need to be performed manually when active document changes, as it does
-        not run callbacks.
-        """
-        self.write(self.read())
-
-
-class FieldWithEditableDefault(Field, Generic[T, F]):
-    def __new__(cls, *args, **kwargs) -> 'FieldWithEditableDefault[T, F]':
-        obj = object.__new__(cls)
-        obj.__init__(*args, **kwargs)
-        return obj
-
-    def __init__(self, field: F, field_with_default: Field[T]):
-        self.field = field
-        self._default_field = field_with_default
-
-        def handle_change_of_default():
-            self.field.default = self._default_field.read()
-        self._default_field.register_callback(handle_change_of_default)
-        handle_change_of_default()
-
-        self.config_group = self.field.config_group
-        self.name = self.field.name
-
-    @property
-    def default(self) -> T:
-        return self.field.default
-
-    @default.setter
-    def default(self, value: T) -> None:
-        self._default_field.write(value)
-
-    def write(self, value: T) -> None:
-        self.field.write(value)
-
-    def read(self) -> T:
-        return self.field.read()
-
-    def register_callback(self, callback: Callable[[], None]) -> None:
-        self.field.register_callback(callback)
-
-    def reset_default(self) -> None:
-        self.field.reset_default()
 
 
 class PieConfig(FieldGroup, Generic[T], ABC):
